hex grid size is now fixed (best screen resolution is 1920x1080).
any lower resolution can cause the svg to extend outside the screen to the right.
propose solution: using an iframe, I think it is appropriate in this particular problem.

it is using a 14x10 som data and hex grid is 14x10 as well.

p.s. The search words are highlighted in both firefox and chrome. Only chrome actually jump to the searched word.